+++ UI/ui_with_protocol_selector_felhantering.py
@@ -2,7 +2,7 @@ import tkinter as tk
 from tkinter import ttk
 from tkinter import messagebox
 from tkinter import filedialog
-from tkinter.constants import NORMAL
+from tkinter.constants import HORIZONTAL, NORMAL
 import replace_values
 import replace_values_qpcr
 import subprocess
@@ -11,7 +11,11 @@ import os.path
 import sys
 import time
 import multiprocessing
+import queue
 import socket
+import math
+
+
 
 # Files and logins for SSH and SCP
 local_user = getlogin() # os.getlogin() get username on local machine
@@ -25,14 +29,6 @@ protocol_qpcr_local_filepath = f'..\\qPCR\\'
 protocol_qpcr_name = 'qpcr_output.py'
 
 font = (16)
-
-
-# Error check to see that the ssh_key is exists.
-if os.path.isfile(key_filename):
-    print("ssh-key read successfully.")
-else:
-    messagebox.showerror('File not found error!', f'SSH Key could not be read. Please check the filepath: {key_filename} and confirm it is placed there')
-    sys.exit(0)
 
 
 class Selector():
@@ -51,8 +47,8 @@ class Selector():
         self.s.configure('my.TLabel', font =('Helvetica', 15))
         self.s.configure('text.TLabel', font=('Helvetica', 10))
 
-        self.label_selection_info = ttk.Label(self.frame, text='Select a protocol', style ='my.TLabel')
-        self.label_selection_info.grid(row=0, column=0, padx=20, pady=20)
+        self.label_selection_info = ttk.Label(self.frame, text='Select a protocol', style ='my.TLabel', anchor="e", justify="right")
+        self.label_selection_info.grid(row=0,column=0,columnspan=2, padx=20, pady=20)
 
         self.button_beads = ttk.Button(self.frame, text='Magnetic beads\nDNA purification', command=self.select_protocol_beads,style='my.TButton')
         self.button_beads.grid(row=1, column=0, padx=10, pady=10, ipadx=5, ipady=1)
@@ -60,8 +56,8 @@ class Selector():
         self.button_qpcr = ttk.Button(self.frame, text='qPCR protocol', command=self.select_protocol_qpcr,style='my.TButton')
         self.button_qpcr.grid(row=1, column=1, padx=10, pady=10, ipadx=10, ipady=10)
 
-        # self.button_test = ttk.Button(self.frame, text='Test', command=self.test,style='my.TButton')
-        # self.button_test.grid(row=1, column=2, padx=10, pady=10,ipadx=10, ipady=10)
+        #self.button_test = ttk.Button(self.frame, text='Test', command=self.test,style='my.TButton')
+        #self.button_test.grid(row=1, column=2, padx=10, pady=10,ipadx=10, ipady=10)
 
     def select_protocol_beads(self):
         '''Closes the frame for protocol selection, but not the root window.
@@ -79,6 +75,12 @@ class Selector():
         self.frame.destroy()
         qPCR_protocol_config()
 
+    def test(self):
+        test_protocol = 'multiprocess_test.py'
+        subprocess.run(f'scp -i {key_filename} {protocol_local_filepath}{test_protocol} {username}@{ip}:{protocol_robot_filepath}{test_protocol}')
+        subprocess.run(f'ssh -i {key_filename} {username}@{ip} -t "sh -lic" \'opentrons_execute {protocol_robot_filepath}{test_protocol}\'')
+        
+
 class Bead_protocol_config():
     '''Contains a frame with widgets used configure a magnetic bead DNA purification protocol.
     The frame will be added to the root window when initialized and destroyed (closed)
@@ -92,11 +94,11 @@ class Bead_protocol_config():
         self.frame.grid()
 
         # Labels
-        self.label_sample_no = ttk.Label(self.frame, text='How many samples: \n(Valid values between 1-96 samples)', style='text.TLabel')
-        self.label_sample_vol = ttk.Label(self.frame, text='Volume sample: \n(Valid values between 15-40 μl)', style='text.TLabel')
+        self.label_sample_no = ttk.Label(self.frame, text='Number of samples: \n(Valid values between 1-96 samples)', style='text.TLabel')
+        self.label_sample_vol = ttk.Label(self.frame, text='Sample volume: \n(Valid values between 15-40 μl)', style='text.TLabel')
         self.label_bead_ratio = ttk.Label(self.frame, text='Bead:Sample ratio: \n(Valid ratios between 0.5-1.5)', style='text.TLabel')
         self.label_ethanol = ttk.Label(self.frame, text='Number of ethanol washes: ', style='text.TLabel')
-        self.label_eb = ttk.Label(self.frame, text='Volume Elution buffer: \n(Valid values between 15-25 μl)', style='text.TLabel')
+        self.label_eb = ttk.Label(self.frame, text='Elution buffer volume: \n(Valid values between 15-25 μl)', style='text.TLabel')
 
         self.label_sample_no.grid(row=0, column=0, padx= 10, pady= 10, sticky=tk.W)
         self.label_sample_vol.grid(row=2, column=0, padx= 10, pady= 10, stick=tk.W)
@@ -145,7 +147,11 @@ class Bead_protocol_config():
 
         self.window = tk.Toplevel()
         sample_no = int(self.entry_sample_no.get())
-        Checkbox(parent=self.window, protocol_type=protocol_dna_name, num_samples=sample_no)
+        sample_vol = float(self.entry_sample_vol.get())
+        ratio = float(self.entry_bead_ratio.get())
+        EB=float(self.entry_eb.get()) 
+        etoh=self.ethanol_var.get()
+        Checkbox(parent=self.window, protocol_type=protocol_dna_name, num_samples=sample_no, sample_vol=sample_vol, ratio=ratio, EB=EB, etoh=etoh)
  
     def ok_button(self):
         ''' Checks if all entries are valid.
@@ -222,13 +228,13 @@ class Bead_protocol_config():
         except ValueError:
             messagebox.showerror('Notice', 'You did not enter valid numbers')
 
+
     # Should not be needed
     # Protocol is started from checklist
     """
     def start_run(self):
         '''Uploads the new protocol using 
         scp -i <key> <file_to_upload> <where_to_place_it>
-
         Then launches the new protocol using
         ssh -i <key> <login> -t "sh -lic" <command>
         -t creates a pseudo terminal on the remote machine (?)
@@ -344,7 +350,6 @@ class qPCR_protocol_config():
         # scp -i <key> <file_to_upload> <where_to_place_it>
         subprocess.run(f'scp -i {key_filename} {protocol_qpcr_local_filepath}{protocol_qpcr_name} {username}@{ip}:{protocol_robot_filepath}{protocol_qpcr_name}')
         print(f'would have run:\nubprocess.run(scp -i {key_filename} {protocol_qpcr_local_filepath}{protocol_qpcr_name} {username}@{ip}:{protocol_robot_filepath}{protocol_qpcr_name}')
-
         # Launch the new protocol using
         # ssh -i <key> <login> <command>
         # -t creates a pseudo terminal on the remote machine (?)
@@ -459,7 +464,8 @@ class Tube_rack_grid():
 
 class Checkbox:
     '''Checkbox class containing checkboxes and other stuff. very bare bones at the moment'''
-    def __init__(self, parent, protocol_type: str, num_samples=None):
+    def __init__(self, parent, protocol_type: str, num_samples=None, sample_vol=None, ratio=None, EB=None, etoh=None):
+        
         self.parent = parent
         self.frame = ttk.Frame(self.parent)
         self.frame.pack()
@@ -471,18 +477,25 @@ class Checkbox:
             self.pipette_text = '\n     Left: P10 single-channel\n     Right: Any'
             self.volumes_label = '4. Fill each tube rack according to its tab.\n    The tabs can be selected on the row above the image.'
         elif self.protocol_type.startswith('dna') and num_samples >= 8: # 8-96 DNA cleaning
+            columns=math.ceil(num_samples/8)
+            beads=sample_vol*ratio*columns+60
+            vol_eb=EB*columns+60
+            vol_etoh=etoh*200+100
             self.protocol = [protocol_local_filepath, protocol_dna_name]
-            self.image_name = '..\\ui\\deck_96.gif'
+            self.image_name = 'ui\\deck_96.gif'
             self.pipette_text = '\n     Left: P10 8-channel\n     Right: P300 8-channel'
-            self.volumes_label = '\n     Fill wells on the ethanol deep well plate\n     corresponding to those that have samples.\n          Ethanol:           200 + ??? extra μl per well\n          Elution buffer: ___  μl per well\n          SPRI beads: __ µl per well'
+            self.volumes_label = '\n     Magnetic beads: '+str(beads)+'µl per well \n     Elution buffer: '+ str(vol_eb)+ ' μl per well\n     EtOH: Fill the wells on the EtOH plate\n               corresponding to the wells with samples;\n               '+ str(vol_etoh)+ ' μl per well'
             self.add_image(self.frame, self.image_name)
         elif self.protocol_type.startswith('dna') and num_samples < 8: # 1-7 DNA cleaning
+            columns=math.ceil(num_samples/8)
+            beads=sample_vol*ratio*columns+60
+            vol_eb=EB*columns+60
+            vol_etoh=etoh*200+100
             self.protocol = [protocol_local_filepath, protocol_dna_name]
-            self.image_name = '..\\ui\\deck_less_8.gif'
+            self.image_name = 'ui\\deck_less_8.gif'
             self.pipette_text = '\n     Left: P10 8-channel\n     Right: P300 8-channel'
-            self.volumes_label = '\n     Fill wells on the ethanol deep well plate \n     corresponding to those that have sample.\n          Ethanol:           200 + ??? extra μl per well\n          Elution buffer: ___  μl per well\n          SPRI beads: __ µl per well'
+            self.volumes_label = '\n     Magnetic beads: '+str(beads)+'µl per well \n     Elution buffer: '+ str(vol_eb)+ ' μl per well\n     One Cleaning: Fill column 5 with 200µl EtOH on the liquids plate \n     Two Cleaning: Fill column 5 & 6 with 200µl EtOH on the liquids plate'
             self.add_image(self.frame, self.image_name)
-
         else:
             messagebox.showerror('Error', f"Invalid protocol type '{self.protocol_type}' entered.")
     
@@ -490,10 +503,10 @@ class Checkbox:
         
         #self.start_button = tk.Button(self.frame, text='Start protocol', command=self.start_protocol, state=tk.DISABLED)
         self.connection_button = ttk.Button(self.frame, text='Check Connection', command= self.check_ssh, style='my.small.TButton')
-        self.connection_button.grid(row=3, column=1, padx=10, pady=10, ipadx=5, ipady=5, sticky=tk.W)
+        self.connection_button.grid(row=3, column=1, padx=10, pady=10, ipadx=5, ipady=5, sticky=tk.E)
         
         self.run_protocol_button = ttk.Button(self.frame, text='Run Protocol', command= self.run_protocol, state='disabled', style='my.small.TButton')
-        self.run_protocol_button.grid(row=20, column= 1, padx=10, pady=10, ipadx=5, ipady=5, sticky=tk.W)
+        self.run_protocol_button.grid(row=20, column= 2, padx=10, pady=10, ipadx=5, ipady=5, sticky=tk.E)
         
         self.label1 = ttk.Label(self.frame, text='1. Check the ssh-connection', font=font).grid(row=0, column=1, sticky=tk.W, padx=20, pady=20, columnspan=2)
         self.label2 = ttk.Label(self.frame, text='2. Check the pipettes:' + self.pipette_text, font=font).grid(row=5, column=1, sticky=tk.W, padx=20, pady=20, columnspan=2)
@@ -501,7 +514,7 @@ class Checkbox:
 
         self.volumes = ttk.Label(self.frame, text=self.volumes_label, font=font).grid(row=11, column=1, sticky=tk.W, padx=20, pady=0, columnspan=2)
 
-        self.connection_status = ttk.Label(self.frame, text='Check Connection', style='my.TLabel', foreground= 'green')
+        self.connection_status = ttk.Label(self.frame, text='   Check Connection', style='my.TLabel', foreground= 'green')
         self.connection_status.grid(row=3, column=2, sticky=tk.W, padx=20, pady=20)
 
     def add_image(self, parent, image_path):
@@ -523,30 +536,55 @@ class Checkbox:
 
 
     def check_ssh(self):
-        '''This function should check if you have a ssh-connection,
-        the host variable should be changed to the robot ip (i think)'''
+        '''This function should check if you have a ssh-connection.
+        Creates a Queue object and passes it to a Process subclass, Threaded_ssh_check().
+        The queue creates a connection between the main process (the UI) and this new process.
+        When the process is started, its run() method is executed,
+        which tries to create a socket connection to the robot ip.
+        After 1 second, calls try_connection() to check if the queue is empty,
+        i.e. if the attempts at creating a socket conneciton is finished.
+        If not, checks again in 3 seconds. 
+        '''
 
-        self.connection_status.config(text='Checking connection...', foreground='black')
+        self.connection_status.config(text='      Checking connection...', foreground='black')
         self.connection_status.update()
-        # host = 'localhost'
-        host = ip
-        port = 22
-
-        try:
-            test_socket = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-            test_socket.connect((host, port))
-        except socket.error as error_msg:
-        # not up, log reason from ex if wanted
-            # tk.messagebox.showerror('Notice', 'Could not establish a ssh-connection')
-            print(error_msg)
-            self.connection_status.config(text='Connection failed', foreground='red') #, wraplength=200)
-        else:
-            test_socket.close()
-            self.run_protocol_button.config(state=tk.NORMAL)
-            self.connection_status.config(text='Connection OK', foreground='green')
-            
         
-            
+        self.valid_connection = False
+
+        self.connection_progress = ttk.Progressbar(self.frame, orient=tk.HORIZONTAL, length=200, mode='indeterminate', )
+        self.connection_progress.grid(row=4, column=2)
+        self.connection_progress.start(10)
+        self.connection_progress.update()
+
+        self.connection_button.config(state=tk.DISABLED)
+
+
+        self.queue = multiprocessing.Queue()
+        self.process = Threaded_ssh_check(self.queue)
+        self.process.start()
+
+        self.connection_progress.after(1000, self.try_connection)
+ 
+        
+    def try_connection(self):
+        try:
+            # valid_connection = True
+            valid_connection = self.queue.get_nowait()
+        except queue.Empty:
+            print('checking queue')
+            self.connection_progress.after(3000, self.try_connection)
+        else:
+            if not valid_connection:
+                self.connection_button.config(state=tk.NORMAL)
+                self.connection_status.config(text='Connection failed', foreground='red') #, wraplength=200)
+                self.connection_progress.destroy()
+                
+            elif valid_connection:
+                self.connection_button.config(state=tk.NORMAL)
+                self.run_protocol_button.config(state=tk.NORMAL)
+                self.connection_status.config(text='Connection OK', foreground='green')
+                self.connection_progress.destroy()
+
         
     def run_protocol(self):
         '''This function runs the protocols, which one it runs is determined by the variable protocol_type when the an object is created.'''
@@ -590,9 +628,9 @@ class Checkbox:
                 log = subprocess.run(f'ssh -i {key_filename} {username}@{ip} -t "sh -lic" \'opentrons_execute {protocol_robot_filepath}{self.protocol[1]}\'', stdout=subprocess.PIPE).stdout.decode('utf-8')
                 print(log)
                 if 'Protocol Complete' in log:
-                    tk.messagebox.showinfo('Protocol Completed', 'Protocol was completed successfully!', parent=self.frame)
+                    tk.messagebox.showinfo('Protocol Completed', 'Protocol was completed successfully!')
                 else:
-                    tk.messagebox.showerror('Protocol Error', 'There was an error in running the protocol.', parent=self.frame)
+                    tk.messagebox.showerror('Protocol Error', 'There was an error in running the protocol.')
             except:
                 messagebox.showerror('Error', 'There was an error running the powershell SSH connect command.')      
             
@@ -613,21 +651,57 @@ class Checkbox:
             print(f'would have run:\nsubprocess.run(ssh -i {key_filename} {username}@{ip} -t "sh -lic" \'opentrons_execute {protocol_robot_filepath}{protocol_qpcr_name}\')')
         '''
 
+class Threaded_ssh_check(multiprocessing.Process):
+    def __init__(self, queue):
+        super().__init__()
+        self.queue = queue
+    
+
+    def run(self):
+        # host = 'localhost'
+        host = ip
+        port = 22
+        try:
+            print(1)
+            test_socket = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
+            print(2)
+            test_socket.settimeout(9)
+            test_socket.connect((host, port))
+            print(3)
+        except (socket.error, socket.timeout) as error_msg:
+            # tk.messagebox.showerror('Notice', 'Could not establish a ssh-connection')
+            print(error_msg)
+            # self.connection_status.config(text='Connection failed', foreground='red') #, wraplength=200)
+            self.queue.put(False)
+        else:
+            print(4)
+            test_socket.close()
+            # self.run_protocol_button.config(state=tk.NORMAL)
+            # self.connection_status.config(text='Connection OK', foreground='green')
+            self.queue.put(True)
+
+
 def run_gui():
     # Creates a root window
     root = tk.Tk()
-    root.title('Protocol selector test')
+    root.title('Protocol selector')
 
     # Creates a frame for the root window with widgets for protocol selection. 
     Selector()
     #Checkbox('qpcr_output.py')
     #Checkbox('dna_cleaning_output.py')
 
+        # Error check to see that the ssh_key is exists.
+    if os.path.isfile(key_filename):
+        print("ssh-key read successfully.")
+    #else:
+        #messagebox.showerror('File not found error!', f'SSH Key could not be found. Please check the filepath: {key_filename} and confirm it is placed there')
+        #sys.exit(0)
+
     root.mainloop()
 
 # Small function to enable multiprocessing later - used only for error-checking.
 def scp_transfer(protocol):
-    test = f'scp -i {key_filename} {protocol[0]}{protocol[1]} {username}@{ip}:{protocol_robot_filepath}{protocol[1]}'
     subprocess.run(f'scp -i {key_filename} {protocol[0]}{protocol[1]} {username}@{ip}:{protocol_robot_filepath}{protocol[1]}')
     return
